+++ src/app/endpoints/streaming_query.py
@@ -103,7 +103,7 @@ from utils.query import (
     validate_attachments_metadata,
     validate_model_provider_override,
 )
-from utils.quota import check_tokens_available, get_available_quotas
+from utils.quota_utils import check_tokens_available, get_available_quotas
 from utils.responses import (
     build_mcp_tool_call_from_arguments_done,
     build_tool_call_summary,
