+++ src/utils/agents/streaming.py
@@ -59,7 +59,7 @@ from utils.agents.tool_processor import (
 from utils.conversations import append_turn_items_to_conversation
 from utils.pydantic_ai import build_agent
 from utils.query import consume_query_tokens, store_query_results
-from utils.quota import get_available_quotas
+from utils.quota_utils import get_available_quotas
 from utils.responses import (
     deduplicate_referenced_documents,
     maybe_get_topic_summary,
